test: uncomment proxy-connect test
